Refactor: Centralize types and logging utilities
Moves common types and tprint functions to dedicated modules.
