+++ src/training/steps/pre_training/feature_lookback_optimization/error_handling/__init__.py
@@ -4,13 +4,8 @@ Error Handling Module.
 This module provides standardized error handling for feature lookback optimization.
 """
 
-# Import tprint for enhanced logging
-try:
-    from src.utils.tprint import tprint_debug
-    TPRINT_AVAILABLE = True
-except ImportError:
-    TPRINT_AVAILABLE = False
-    def tprint_debug(*args, **kwargs): print("DEBUG:", *args, **kwargs)
+# Import centralized tprint utilities
+from ..utils.tprint_utils import tprint_debug, TPRINT_AVAILABLE
 
 from .error_handler import StandardizedErrorHandler, ErrorSeverity, ErrorCategory, ErrorDetails
 # ErrorRecoveryResult doesn't exist - removed from import
@@ -23,6 +18,5 @@ __all__ = [
     'StandardizedErrorHandler',
     'ErrorSeverity',
     'ErrorCategory',
-    'ErrorDetails',
-    'ErrorRecoveryResult'
+    'ErrorDetails'
 ]
