+++ src/training/steps/pre_training/feature_lookback_optimization/utils/fast_failing_validation.py
@@ -16,22 +16,12 @@ from enum import Enum
 import time
 import traceback
 
-# Import tprint utilities for enhanced logging
-try:
-    from src.utils.tprint import (
-        tprint, tprint_debug, tprint_info, tprint_warning, 
-        tprint_error, tprint_success, tprint_performance
-    )
-    TPRINT_AVAILABLE = True
-except ImportError:
-    TPRINT_AVAILABLE = False
-    def tprint(*args, **kwargs): print("TPRINT:", *args, **kwargs)
-    def tprint_debug(*args, **kwargs): print("DEBUG:", *args, **kwargs)
-    def tprint_info(*args, **kwargs): print("INFO:", *args, **kwargs)
-    def tprint_warning(*args, **kwargs): print("WARNING:", *args, **kwargs)
-    def tprint_error(*args, **kwargs): print("ERROR:", *args, **kwargs)
-    def tprint_success(*args, **kwargs): print("SUCCESS:", *args, **kwargs)
-    def tprint_performance(*args, **kwargs): print("PERF:", *args, **kwargs)
+# Import centralized tprint utilities
+from .tprint_utils import (
+    tprint, tprint_debug, tprint_info, tprint_warning, 
+    tprint_error, tprint_success, tprint_performance,
+    TPRINT_AVAILABLE
+)
 
 from .error_handling import DataValidationError, OptimizationError
 from .nan_handling import SafeNaNHandler
